refactor: useSignUpFunnel 구현 (#ATR-147)
