+++ apps/service/widgets/sign-up-container/api/index.ts
@@ -1,0 +1,1 @@
+export { default as postSignUpForm } from './postSignUpForm'
