+++ apps/service/widgets/sign-up-container/lib/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useSignUpFunnel } from './useSignUpFunnel'
